requirements.txt
fastapi
uvicorn
requests
transformers
torch 
pillow
openai-whisper
python-multipart
protobuf
sentencepiece